// src/types/api/playListItem.d.ts
import { userProfile } from "@/types/user/user"

export type playListItem = {
    subscribers: [],
    subscribed: boolean,
    creator: userProfile,
    artists: null,
    tracks: null,
    top: boolean,
    updateFrequency: null,
    backgroundCoverId: number,
    backgroundCoverUrl: null,
    titleImage: number,
    titleImageUrl: null,
    englishTitle: null,
    opRecommend: boolean,
    recommendInfo: null,
    subscribedCount: number,
    cloudTrackCount: number,
    userId: number,
    totalDuration: number,
    coverImgId: number,
    privacy: number,
    trackUpdateTime: number,
    trackCount: number,
    updateTime: number,
    commentThreadId: string,
    coverImgUrl:string,
    specialType: number,
    anonimous: boolean,
    createTime: number,
    highQuality: boolean,
    newImported: boolean,
    trackNumberUpdateTime: number,
    playCount: number,
    adType: number,
    description: string,
    tags: string[],
    ordered: true,
    status: number,
    name:string,
    id: number,
    coverImgId_str: string,
    sharedUsers: null,
    shareStatus: null,
    copied: boolean,
    containsTracks: boolean
}